+++ sdk/monitor/azure-monitor-query/src/main/java/com/azure/monitor/query/implementation/metrics/models/ErrorContract.java
@@ -5,7 +5,11 @@
 package com.azure.monitor.query.implementation.metrics.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
 /**
  * Error Response
@@ -14,13 +18,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * follows the OData error response format.).
  */
 @Fluent
-public final class ErrorContract {
+public final class ErrorContract implements JsonSerializable<ErrorContract> {
     /*
      * Error Response
      * 
      * The error object.
      */
-    @JsonProperty(value = "error")
     private ErrorResponse error;
 
     /**
@@ -51,5 +54,38 @@ public final class ErrorContract {
     public ErrorContract setError(ErrorResponse error) {
         this.error = error;
         return this;
+    }
+
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeJsonField("error", this.error);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of ErrorContract from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of ErrorContract if the JsonReader was pointing to an instance of it, or null if it was
+     * pointing to JSON null.
+     * @throws IOException If an error occurs while reading the ErrorContract.
+     */
+    public static ErrorContract fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            ErrorContract deserializedErrorContract = new ErrorContract();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("error".equals(fieldName)) {
+                    deserializedErrorContract.error = ErrorResponse.fromJson(reader);
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedErrorContract;
+        });
     }
 }