+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/OpenAIServiceVersion.java
@@ -38,7 +38,12 @@ public enum OpenAIServiceVersion implements ServiceVersion {
     /**
      * Enum value 2024-03-01-preview.
      */
-    V2024_03_01_PREVIEW("2024-03-01-preview");
+    V2024_03_01_PREVIEW("2024-03-01-preview"),
+
+    /**
+     * Enum value 2024-04-01-preview.
+     */
+    V2024_04_01_PREVIEW("2024-04-01-preview");
 
     private final String version;
 
@@ -60,6 +65,6 @@ public enum OpenAIServiceVersion implements ServiceVersion {
      * @return The latest {@link OpenAIServiceVersion}.
      */
     public static OpenAIServiceVersion getLatest() {
-        return V2024_03_01_PREVIEW;
+        return V2024_04_01_PREVIEW;
     }
 }