+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/models/AzureChatExtensionsMessageContext.java
@@ -24,11 +24,10 @@ public final class AzureChatExtensionsMessageContext {
     }
 
     /*
-     * The contextual information associated with the Azure chat extensions used for a chat completions request.
-     * These messages describe the data source retrievals, plugin invocations, and other intermediate steps taken in
-     * the
-     * course of generating a chat completions response that was augmented by capabilities from Azure OpenAI chat
-     * extensions.
+     *   The contextual information associated with the Azure chat extensions used for a chat completions request.
+     *   These messages describe the data source retrievals, plugin invocations, and other intermediate steps taken in the
+     *   course of generating a chat completions response that was augmented by capabilities from Azure OpenAI chat
+     *   extensions.
      */
     @Generated
     @JsonProperty(value = "citations")
@@ -42,10 +41,9 @@ public final class AzureChatExtensionsMessageContext {
     private String intent;
 
     /**
-     * Get the citations property: The contextual information associated with the Azure chat extensions used for a
-     * chat completions request.
-     * These messages describe the data source retrievals, plugin invocations, and other intermediate steps taken in
-     * the
+     * Get the citations property: The contextual information associated with the Azure chat extensions used for a chat
+     * completions request.
+     * These messages describe the data source retrievals, plugin invocations, and other intermediate steps taken in the
      * course of generating a chat completions response that was augmented by capabilities from Azure OpenAI chat
      * extensions.
      *