+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -1,12 +1,13 @@
 # Release History
 
-## 2.38.0-beta.1 (Unreleased)
+## 2.39.0-beta.1 (Unreleased)
 
 ### Features Added
 
-### Breaking Changes
+- Supported disabling public network access in `Disk` via `disablePublicNetworkAccess()`, for private link feature.
+- Supported disabling public network access in `Snapshot` via `disablePublicNetworkAccess()`, for private link feature.
 
-### Bugs Fixed
+## 2.38.0 (2024-04-16)
 
 ### Other Changes